move handler logic
