+++ packages/snaps-controllers/src/interface/utils.ts
@@ -1,4 +1,5 @@
-import { Component, NodeType } from '@metamask/snaps-ui';
+import type { Component } from '@metamask/snaps-sdk';
+import { NodeType } from '@metamask/snaps-sdk';
 
 export type ComponentState = Record<
   string,
