+++ packages/snaps-utils/src/handler-types.ts
@@ -8,6 +8,7 @@ export enum HandlerType {
   OnNameLookup = 'onNameLookup',
   OnKeyringRequest = 'onKeyringRequest',
   OnHomePage = 'onHomePage',
+  OnUserInput = 'onUserInput',
 }
 
 export type SnapHandler = {
